requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
SQLAlchemy==1.3.22
Werkzeug==1.0.1